+++ spec/factories/pages.rb
@@ -21,8 +21,8 @@ FactoryGirl.define do
       body  nil
     end
 
-    trait :not_found do
-      status_code 404
+    trait :gone do
+      status_code 410
 
       theme nil
       title nil