+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 See README.md before updating this file.
 
 ## Unreleased [#](https://git.cashnetusa.com/trogdor/landable/compare/v1.4.2...master)
+* page#status_code switch 404s to 410s [#104] 
 * Adding Preview Message [#102]
 * Adding Partials to Templates Support [#103]